refactor organization modals
