+++ src/client/src/containers/PageReviewAndGenerate/index.tsx
@@ -12,6 +12,8 @@ import { ThunkDispatch } from "redux-thunk";
 import { AppState } from "../../reducers";
 import RootAction from "../../actions/ActionType";
 import messages from "./messages";
+import AddPagesModal from "./AddPagesModal";
+import PostGenerationModal from "./PostGenerationModal";
 
 interface IDispatchProps {
   openViewLicensesModal: () => any;
@@ -29,6 +31,8 @@ const ReviewAndGenerate = (props: Props) => {
 
   return (
     <div className={styles.container}>
+      <AddPagesModal/>
+      <PostGenerationModal/>
       <div className={styles.reviewContextContainer}>
         <div className={styles.selectionContainer}>
           <h1>{formatMessage(messages.launchYourProject)}</h1>
