+++ src/client/src/containers/ServicesList/index.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import { injectIntl, InjectedIntlProps } from "react-intl";
 import messages from "./strings";
 import styles from "./styles.module.css";
-import CosmosDBSelection from "../CosmosDBSelection";
+import CosmosDBSelection from "./CosmosDBSelection";
 import AppServiceSelection from "../AppServiceSelection";
 import { ServiceState } from "../../reducers/wizardSelectionReducers/services";
 import _ from "lodash";
