+++ src/client/src/translations/whitelist_en.json
@@ -25,15 +25,6 @@
   "azureFunctionsModal.appNameSubLabel",
   "azureFunctionsModal.numFunctionsSubLabel",
   "azureFunctionsSelection.duplicateName",
-  "header.signOut",
-  "azureLoginModal.azureReadMore",
-  "azureLoginModal.signIn",
-  "azureLoginModal.createAccount",
-  "azureLoginModal.freeAccountAnswer",
-  "azureLoginModal.freeAccountQuestion",
-  "azureLoginModal.getStartedWithAzure",
-  "azureLoginModal.paymentAnswer",
-  "azureLoginModal.paymentQuestion",
   "azureStudent.azureForStudent",
   "azureStudent.noCreditCard",
   "azureStudent.startStudentAccount",
@@ -68,10 +59,23 @@
   "footer.generate",
   "footer.navAriaLabel",
   "licenses.redirectLinkLabel",
-  "newProject.header",
-  "newProject.body",
   "selectPages.pagesTitleQuestion",
   "selectableCard.details",
+  "azureLoginModal.azureReadMore",
+  "azureLoginModal.signIn",
+  "azureLoginModal.createAccount",
+  "azureLoginModal.freeAccountAnswer",
+  "azureLoginModal.freeAccountQuestion",
+  "azureLoginModal.getStartedWithAzure",
+  "azureLoginModal.paymentAnswer",
+  "azureLoginModal.paymentQuestion",
+  "header.signOut",
+  "newProject.header",
+  "newProject.body",
+  "licenses.viewLicenses",
+  "instructionHeading.launchYourProject",
+  "context.almostDone",
+  "context.quickStart",
   "postGenerationModal.restartWizard",
   "postGenerationModal.deploymentFailure",
   "postGenerationModal.closeWizard",
@@ -92,10 +96,6 @@
   "redirectModal.toContinue",
   "redirectModal.privacyStatement",
   "redirectModal.OK",
-  "licenses.viewLicenses",
-  "instructionHeading.launchYourProject",
-  "context.almostDone",
-  "context.quickStart",
   "rightSidebar.yourProjectDetails",
   "rightSidebar.projectName",
   "rightSidebar.location",
